Rename ChatRequest to ChatUri which is more accurate and move boilerplate URI creation logic from test to ChatUri.
