+++ guides/micronaut-websocket/java/src/main/java/example/micronaut/ChatClientConnect.java
@@ -6,8 +6,6 @@ import jakarta.inject.Singleton;
 import org.reactivestreams.Publisher;
 import reactor.core.publisher.Flux;
 
-import java.net.URI;
-
 @Singleton // <1>
 public class ChatClientConnect {
     private final ApplicationContext appContext;
@@ -16,9 +14,9 @@ public class ChatClientConnect {
         this.appContext = appContext;
     }
 
-    public ChatClientEndpoint connect(URI uri) { // <3>
+    public ChatClientEndpoint connect(ChatUri uri) { // <3>
         WebSocketClient webSocketClient = appContext.getBean(WebSocketClient.class);
-        Publisher<ChatClientEndpoint> client = webSocketClient.connect(ChatClientEndpoint.class, uri);
+        Publisher<ChatClientEndpoint> client = webSocketClient.connect(ChatClientEndpoint.class, uri.delegate());
 
         return Flux.from(client).blockFirst();
     }
