+++ guides/micronaut-websocket/micronaut-websocket.adoc
@@ -29,7 +29,7 @@ source:ChatClientEndpoint[]
 <4> Binds to the WebSocket "onMessage" event
 <5> Add messages to our history
 <6> Binds to the WebSocket "onClose" event
-<7> Any method name that start with (or equals) "broadcast" or "send" will be used to broadcast messages as long as the parameter is a String or a bean.
+<7> Any method name that start with (or equals) "broadcast" or "send" will be used to broadcast messages as long as the parameter is a String or a bean
 
 Next, we'll create a connection factory to instantiate chat clients:
 source:ChatClientConnect[]
@@ -38,7 +38,7 @@ source:ChatClientConnect[]
 <3> Create the WebSocket client for the given URI specifying the username and topic
 
 This bean will help in constructing URIs with parameters:
-source:ChatRequest[]
+source:ChatUri[]
 <1> We will create maps from this bean so it needs to be @Introspected
 <2> Username of the chat client
 <3> Topic of the chat client
